splitter template route komponentene inn i flere filer
