+++ skribenten-web/frontend/src/routes/saksnummer_/$saksId/brevvelger/$templateId/-FavoriteButton.tsx
@@ -1,0 +1,54 @@
+import { css } from "@emotion/react";
+import { StarFillIcon, StarIcon } from "@navikt/aksel-icons";
+import { Button } from "@navikt/ds-react";
+import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
+
+import { addFavoritt, deleteFavoritt, getFavoritter } from "~/api/skribenten-api-endpoints";
+
+import { Route } from "./route";
+
+export default function FavoriteButton() {
+  const { templateId } = Route.useParams();
+  const queryClient = useQueryClient();
+  const isFavoritt = useQuery({
+    ...getFavoritter,
+    select: (favoritter) => favoritter.includes(templateId),
+  }).data;
+
+  const toggleFavoritesMutation = useMutation<unknown, unknown, string>({
+    mutationFn: (id) => (isFavoritt ? deleteFavoritt(id) : addFavoritt(id)),
+    onSettled: () => queryClient.invalidateQueries({ queryKey: getFavoritter.queryKey }),
+  });
+
+  if (isFavoritt) {
+    return (
+      <Button
+        css={css`
+          width: fit-content;
+        `}
+        data-cy="remove-favorite-button"
+        icon={<StarFillIcon aria-hidden />}
+        onClick={() => toggleFavoritesMutation.mutate(templateId)}
+        size="small"
+        variant="secondary"
+      >
+        Fjern som favoritt
+      </Button>
+    );
+  }
+
+  return (
+    <Button
+      css={css`
+        width: fit-content;
+      `}
+      data-cy="add-favorite-button"
+      icon={<StarIcon aria-hidden />}
+      onClick={() => toggleFavoritesMutation.mutate(templateId)}
+      size="small"
+      variant="secondary-neutral"
+    >
+      Legg til som favoritt
+    </Button>
+  );
+}
