+++ skribenten-web/frontend/src/routes/saksnummer_/$saksId/brevvelger/$templateId/-SelectSensitivity.tsx
@@ -1,0 +1,35 @@
+import { css } from "@emotion/react";
+import { Radio, RadioGroup } from "@navikt/ds-react";
+import { Controller } from "react-hook-form";
+
+function SelectSensitivity() {
+  return (
+    <Controller
+      name="isSensitive"
+      render={({ field, fieldState }) => (
+        <RadioGroup
+          data-cy="is-sensitive"
+          legend="Er brevet sensitivt?"
+          {...field}
+          css={css`
+            .navds-radio-buttons {
+              margin: 0 !important;
+            }
+
+            .navds-radio__label {
+              padding: var(--a-spacing-2) 0;
+            }
+          `}
+          error={fieldState.error?.message}
+          size="medium"
+          value={field.value ?? null}
+        >
+          <Radio value>Ja</Radio>
+          <Radio value={false}>Nei</Radio>
+        </RadioGroup>
+      )}
+    />
+  );
+}
+
+export default SelectSensitivity;
