+++ skribenten-web/frontend/tsr.config.json
@@ -1,0 +1,5 @@
+{
+    "routeFileIgnorePrefix": "-",
+    "routesDirectory": "./src/routes",
+    "generatedRouteTree": "./src/routeTree.gen.ts"
+}
